Add support for include directives.
These must be on a line of their own and take the form
>> file name <<
the specified file name (considered relative to the file containing
the include directive) will be parsed and its chunks inserted into
the parse result in place of the include directive. The Def chunks
now keep track of the source file name as well as the source line,
as there can now be multiple source files for a single parse tree.